+++ CHANGELOG.md
@@ -4,6 +4,7 @@ BUG FIXES:
   * api: Fix TLS in logs and fs commands/APIs [GH-2290]
   * cli/plan: Fix diff alignment and remove no change DC output [GH-2465]
   * client: Fix panic when restarting non-running tasks [GH-2480]
+  * client: Fix env vars when multiple tasks and ports present [GH-2491]
 
 ## 0.5.5 (March 14, 2017)